+++ frontend/src/components/blocks/BlockView.tsx
@@ -7,6 +7,7 @@ import { ChartBlockView } from './ChartBlock'
 import { CodeBlockView } from './CodeBlock'
 import { DiagramBlockView } from './DiagramBlock'
 import { SlidesBlockView } from './SlidesBlock'
+import { EditsBlockView } from './EditsBlock'
 import { MapBlockView } from './MapBlock'
 import { SourcesBlockView } from './SourcesBlock'
 
@@ -27,6 +28,8 @@ export function BlockView({ block }: { block: Block }) {
       return <DiagramBlockView block={block} />
     case 'slides':
       return <SlidesBlockView block={block} />
+    case 'edits':
+      return <EditsBlockView block={block} />
     case 'map':
       return <MapBlockView block={block} />
     case 'sources':
@@ -61,6 +64,8 @@ export function BlockSkeleton({ blockType }: { blockType: string }) {
       return <Skeleton className="h-56 w-full" />
     case 'slides':
       return <Skeleton className="aspect-[16/9] w-full" />
+    case 'edits':
+      return <Skeleton className="h-24 w-full" />
     case 'code':
       return <Skeleton className="h-24 w-full" />
     default: